refactor(storage): split queries.rs into domain modules; remove dead code
Split src/storage/queries.rs (5907 lines) into 6 domain-specific modules:
- chat.rs: chat, message, session, channel-log queries + tests
- tool.rs: tool call and LLM usage queries + tests
- sleep.rs: sleep run, step, memory snapshot queries + tests
- pulse.rs: pulse run queries + tests
- episode.rs: episode event and rollup queries + tests
- mod.rs: type definitions, DB factory, macros

Remove 14 pub(crate) methods unused by any runtime code:
sleep: update_sleep_run_skipped, commit_step_success, get_sleep_run_step,
  upsert_sleep_checkpoint, list_sleep_checkpoints, create_memory_snapshot,
  update_memory_snapshot_after, get_snapshots_for_agent, get_latest_snapshot_for_file
episode: insert_episode_events, list_episode_events, count_episode_events,
  get_episode_rollup, list_episode_rollups_in_range

Also remove 30 corresponding unit tests and their helpers.
All remaining 1,234 tests pass; clippy clean.
